+++ src/packages/piral-axios/src/types.ts
@@ -1,0 +1,15 @@
+import 'piral-core';
+import { AxiosInstance, AxiosRequestConfig } from 'axios';
+
+declare module 'piral-core/lib/types/custom' {
+  interface PiletCustomApi extends PiletAxiosApi {}
+}
+
+export type AxiosConfig = AxiosRequestConfig;
+
+export interface PiletAxiosApi {
+  /**
+   * Gets the associated axios instance.
+   */
+  axios: AxiosInstance;
+}