+++ src/pages/www/src/scripts/components/Capabilities.tsx
@@ -2,13 +2,13 @@ import * as React from 'react';
 import { Capability } from './Capability';
 
 export const Capabilities: React.FC = () => (
-  <div className="container my-5 py-2">
-    <h2 className="text-center font-weight-bold my-5">Breaking the Frontend Monolith!</h2>
-    <div className="text-center">
+  <div className="container container-section text-center">
+    <h2>Breaking the Frontend Monolith!</h2>
+    <p>
       Piral is a framework for next generation portal applications. It allows you to build web apps that follow the
       microfrontends architecture in basically no time. Piral lets you use your standard tooling to be as efficient as
       possible.
-    </div>
+    </p>
     <div className="row">
       <Capability image={require('../../assets/capability-cloud.png')} title="Cloud Ready">
         Perfect for distributed systems running in the cloud.